Create a cli called 'devj'
This separates out a cli from the main repo.
The cli imports the new `journal` package and uses it to do the same
thing that it previously did.
However, it is now necessary to add the command `new` after the cli,
e.g. `devj new` in order to have the same effect.

Later, I should add a message or functionality for when no arguments or
flags are given.