Add edit command to edit most recent entry
Currently there is a bug where edit will not grab the current entry for
the day, but only an entry older than that.
Also, it should probably create an entry for the day if one does not
exist, though perhaps that should itself be another command, like
`newedit` or `nedit` or something.