Add DateString, LatestEntry, fix edit bug
The edit command used to only be able to get any entry prior
to today.
However, this was because the PreviousEntry function worked as
intended.
Updating it to LatestEntry, which includes today, fixes it.

Also, due to the change to LatestEntry, the MakeNewEntry
function would create an empty file prior to looking for a
file to choose, causing it to always create a blank file.
This is also fixed.

Finally, a helper function DateString has been added, which
gets around the several-times-copied code of getting the
current date and creating a string from that.